fixed compile errors and updated sdm lib dependency
