+++ sdm_sensor.cpp
@@ -1,5 +1,6 @@
 #include "sdm_sensor.h"
 #include "esphome/core/log.h"
+#include "SDM.h"
 
 namespace esphome {
 namespace sdm_sensor {
@@ -7,9 +8,9 @@ namespace sdm_sensor {
 static const char *TAG = "sdm_sensor.SDMSensor";
 
 float SDMSensor::readVal(uint16_t reg) {
-	float val = sdm.readVal(reg, channel_);
+	float val = sdm->readVal(reg, channel_);
 
-	uint16_t err = sdm.getErrCode(true);
+	uint16_t err = sdm->getErrCode(true);
 	if (err != 0) {
 		ESP_LOGE(TAG, "Error %i", err);
 	}
@@ -24,10 +25,10 @@ void SDMSensor::setup() {
 void SDMSensor::update() {
 	ESP_LOGD(TAG, "Start reading");
 
-	INODE_READ_PUBLISH(this->sdm120ct_voltage, 				SDM120CT_VOLTAGE);
-	INODE_READ_PUBLISH(this->sdm120ct_current, 				SDM120CT_CURRENT);
-	INODE_READ_PUBLISH(this->sdm120ct_power, 				SDM120CT_POWER);
-	INODE_READ_PUBLISH(this->sdm120ct_import_active_energy, SDM120CT_IMPORT_ACTIVE_ENERGY);
+	INODE_READ_PUBLISH(this->sdm120ct_voltage, 				SDM_PHASE_1_VOLTAGE);
+	INODE_READ_PUBLISH(this->sdm120ct_current, 				SDM_PHASE_1_CURRENT);
+	INODE_READ_PUBLISH(this->sdm120ct_power, 				SDM_PHASE_1_POWER);
+	INODE_READ_PUBLISH(this->sdm120ct_import_active_energy, SDM_IMPORT_ACTIVE_ENERGY);
 
 	ESP_LOGD(TAG, "End reading");
 }
@@ -40,10 +41,10 @@ void SDMSensor::dump_config() {
 	ESP_LOGD(TAG, "RX pin: %i", rx_pin_);
 	ESP_LOGD(TAG, "TX pin: %i", tx_pin_);
 
-	LOG_SENSOR("  ", "sdm120ct_voltage", this->sdm120ct_voltage);
-	LOG_SENSOR("  ", "sdm120ct_current", this->sdm120ct_current);
-	LOG_SENSOR("  ", "sdm120ct_power", this->sdm120ct_power);
-	LOG_SENSOR("  ", "sdm120ct_import_active_energy", this->sdm120ct_import_active_energy);
+	LOG_SENSOR("  ", "sdm120ct_voltage", this->sdm120ct_voltage_);
+	LOG_SENSOR("  ", "sdm120ct_current", this->sdm120ct_current_);
+	LOG_SENSOR("  ", "sdm120ct_power", this->sdm120ct_power_);
+	LOG_SENSOR("  ", "sdm120ct_import_active_energy", this->sdm120ct_import_active_energy_);
 }
 
 } //namespace sdm_sensor
